resolução dos exercicios de condicionais
resolução dos exercicios de condicionais.
resolução dos 1 ao 7.

+++ Sintaxe com Python/exercicios_de_operadores.py
@@ -75,7 +75,7 @@ elif numero_1 != numero_2:
 
 numero_1 = int(input("Digite um numero: "))
 operador = input("Qual operação que fazer [+], [-], [*], [/]: ")
-numero_2 = float(input("Digite um segundo numero: "))
+numero_2 = int(input("Digite um segundo numero: "))
 
 if operador == "+":
     numero_1 += numero_2
